bg: fix-product-edit-notification This Commit has the changes that organizes the codes by splint the re usable components out of the entire long file
