+++ src/components/dashboards/admin/contents/Products.jsx
@@ -6,17 +6,17 @@ import { toast } from 'react-toastify';
 import { NavLink } from 'react-router-dom';
 import Swal from 'sweetalert2';
 import BreadCumb from '../../../BreadCumb';
-import FormInput from '../../../formControlscomponents/formInput/FormInput';
-import imgInputController from '../../../formControlscomponents/formInput/imgInputController';
 import Loading from '../../../Loading';
+
+import Modal from '../../../Modal';
 
 /* eslint import/no-webpack-loader-syntax: off */
 import {
   deleteSellerProducts,
   fetchSellerProducts,
-  updateSellerProducts,
-  updateProductAttribute,
 } from '../../../../redux/reducers/seller/SellerProductSlice';
+import FullProduct from '../../../FullProduct';
+import FullProductAttribute from '../../../FullProductAttribute';
 
 const Products = () => {
   const ITEMS_PER_PAGE = 10;
@@ -31,25 +31,11 @@ const Products = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [modalProduct, setModalProduct] = useState(null);
 
-  const { sellerProducts, searchedProducts } = useSelector(
-    (state) => state.sellerProducts
-  );
-
-  const [productFormValue, setProductFormValue] = useState({});
-
-  const [productImg, setproductImg] = useState('');
-  const [newProductImg, setNewProductImg] = useState(null);
-
-  const [updateProductbtn, setUpdateProductbtn] = useState({
-    text: 'Update',
-    loading: 'hidden',
-    disabled: false,
-  });
+  const { sellerProducts } = useSelector((state) => state.sellerProducts);
   const { products, status, error } = sellerProducts;
   const data = products?.data;
 
-  const { updateProductStatus, updateProductAttributeStatus, deleteStatus } =
-    sellerProducts;
+  const { deleteStatus } = sellerProducts;
 
   const handleChangePage = (page) => setCurrentPage(page);
 
@@ -75,103 +61,20 @@ const Products = () => {
     });
   };
 
-  const handleEditProduct = (product) => {
+  const handleCloseModal = () => {
+    setIsModalOpen(false);
+    setModalProduct(null);
+  };
+  const handleOpenModal = () => {
     setIsModalOpen(true);
+  };
+
+  const handleEditProduct = (product) => {
+    handleOpenModal();
     setModalProduct(product);
-    setproductImg(product.image);
-    setProductFormValue(product);
-  };
-
-  const handleTyping = (event) => {
-    setProductFormValue({
-      ...productFormValue,
-      [event.target.name]: event.target.value,
-    });
-  };
-
-  // Handle product attribute
-  const handleTypingAttribute = (event, productAttibuteId) => {
-    const newAttributes = productFormValue.ProductAttributes.map(
-      (attribute) => {
-        if (attribute.id === productAttibuteId)
-          return { ...attribute, [event.target.name]: event.target.value };
-        return attribute;
-      }
-    );
-    setProductFormValue({
-      ...productFormValue,
-      ProductAttributes: newAttributes,
-    });
-  };
-
-  // Handle product Update event
-  const handleUpdateProduct = async (event) => {
-    event.preventDefault();
-    const formElement = event.target;
-    const fromData = new FormData(formElement);
-    if (newProductImg) {
-      fromData.append('image', event.target[0].files[0]);
-    }
-    const formDataObject = Object.fromEntries(fromData.entries());
-    setProductFormValue(formDataObject);
-
-    await dispatch(
-      await updateSellerProducts({ id: formDataObject.id, body: fromData })
-    );
-    if (status === 'succeeded' && updateProductStatus) {
-      toast.success(updateProductStatus, { theme: 'colored' });
-    } else toast.error(error, { theme: 'colored' });
-    await dispatch(fetchSellerProducts());
-  };
-
-  // Handle product attribute update
-  const handleUpdateProductAttribute = async (event) => {
-    event.preventDefault();
-    const formData = new FormData(event.target);
-    if (event.target?.[0]?.files.length) {
-      formData.append('attrImage', event.target[0].files[0]);
-    }
-    const formDataObject = Object.fromEntries(formData.entries());
-    await dispatch(
-      await updateProductAttribute({ id: formDataObject.id, body: formData })
-    );
-
-    if (status === 'succeeded' && updateProductAttributeStatus) {
-      toast.success(updateProductAttributeStatus, { theme: 'colored' });
-    } else toast.error(error, { theme: 'colored' });
-    await dispatch(fetchSellerProducts());
-  };
-
-  // On product image  change
-  const imgDisplay = (e) => {
-    setproductImg(window.URL.createObjectURL(e.target.files[0]));
-    setNewProductImg(window.URL.createObjectURL(e.target.files[0]));
-  };
-
-  // On product image attribute change
-  const imgAttributeDisplay = (event) => {
-    const formElement = event.target
-      .closest('form')
-      .querySelector('input[name="id"]');
-    const newAttributes = productFormValue.ProductAttributes.map(
-      (attribute) => {
-        if (Number(attribute.id) === Number(formElement.value)) {
-          return {
-            ...attribute,
-            attrImage: window.URL.createObjectURL(event.target.files[0]),
-          };
-        }
-        return attribute;
-      }
-    );
-    setProductFormValue({
-      ...productFormValue,
-      ProductAttributes: newAttributes,
-    });
   };
 
   // filtred products
-
   const filteredProducts = data?.products?.filter(
     (product) =>
       product.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
@@ -189,11 +92,11 @@ const Products = () => {
   const handleSearchChange = (event) => {
     setSearchTerm(event.target.value);
   };
+
   useEffect(() => {
     const handleOutsideClick = (event) => {
       if (!event.target.closest('.modal-content')) {
-        setIsModalOpen(false);
-        setModalProduct(null);
+        handleCloseModal();
       }
     };
     if (isModalOpen) {
@@ -209,10 +112,6 @@ const Products = () => {
   useEffect(() => {
     dispatch(fetchSellerProducts());
   }, []);
-
-  useEffect(() => {
-    setProductFormValue(modalProduct);
-  }, [modalProduct]);
 
   const crumbs = [{ text: 'Products', path: '/dashboard/products' }];
 
@@ -274,11 +173,11 @@ const Products = () => {
           <table className="table-auto w-full">
             <thead className="text-darkBlueColor">
               <tr>
-                <th className="border px-4 py-2"> Image </th>
-                <th className="border px-4 py-2">{t('pname')}</th>
-                <th className="border px-4 py-2">{t('model')}</th>
-                <th className="border px-4 py-2">{t('status')}</th>
-                <th className="border px-4 py-2">{t('option')}</th>
+                <th className="table-header"> Image </th>
+                <th className="table-header">{t('pname')}</th>
+                <th className="table-header">{t('model')}</th>
+                <th className="table-header">{t('status')}</th>
+                <th className="table-header">{t('option')}</th>
               </tr>
             </thead>
             <tbody>
@@ -366,328 +265,10 @@ const Products = () => {
         )}
       </div>
 
-      <div
-        className={`${
-          isModalOpen && modalProduct ? 'fixed' : 'hidden'
-        } fixed z-5 inset-0 overflow-y-auto bg-opacity-50 bg-black`}
-      >
-        <div className="relative z-[9999] max-h-screen overflow-y-auto top-1/2 transform -translate-y-1/2 w-11/12 mx-auto max-w-2xl">
-          <div className="modal-content sticky bg-white z-[9999] shadow-lg rounded-lg p-8">
-            <div className="flex justify-between items-center mb-10">
-              <h2 className="text-xl font-bold">{t('edit_product')}</h2>
-              <button
-                onClick={() => {
-                  setIsModalOpen(false);
-                  setModalProduct(null);
-                }}
-                type="button"
-                className="text-gray-900 hover:text-gray-600 focus:outline-none"
-                aria-label="Close modal"
-              >
-                <svg className="w-6 h-6 fill-current" viewBox="0 0 20 20">
-                  <path
-                    fillRule="evenodd"
-                    d="M10 8.586L6.879 5.464 5.464 6.879 8.586 10l-2.122 2.121 1.415 1.415L10 11.414l3.121 3.122 1.415-1.415L11.414 10l2.122-2.121-1.415-1.415z"
-                  />
-                </svg>
-              </button>
-            </div>
-            <div>
-              {/* start Prodcut */}
-              <div className="product z-[99] mb-4 border-2 rounded p-3">
-                <h1 className="z-[9999] font-mono text-lg ">{t('product')} </h1>
-                <hr />
-                <form
-                  className=" flex justify-evenly flex-wrap flex-row-reverse"
-                  onSubmit={handleUpdateProduct}
-                >
-                  <div className="w-[60%] md:max-h-[300px] max-h-fit my-5 relative">
-                    {imgInputController(
-                      imgDisplay,
-                      productImg,
-                      false,
-                      t('product_image')
-                    )}
-                  </div>
-                  <div className="w-[100%] min-w-[20em] flex  flex-wrap  ">
-                    <div className="sm:px-3 p-2 m-2">
-                      <label
-                        htmlFor="name"
-                        className=" text-[#2b2b2b] text-sm block"
-                      >
-                        {t('name')}
-                      </label>
-                      <FormInput
-                        id="name"
-                        onChange={(event) => handleTyping(event)}
-                        name="name"
-                        value={productFormValue && productFormValue.name}
-                        type="text"
-                        required
-                        className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                      />
-                    </div>
-
-                    <div className="sm:px-3 p-2 mx-2">
-                      <label
-                        htmlFor="model"
-                        className=" text-[#2b2b2b] text-sm block "
-                      >
-                        {t('model')}
-                      </label>
-                      <FormInput
-                        id="model"
-                        onChange={(event) => handleTyping(event)}
-                        name="model"
-                        value={productFormValue && productFormValue.model}
-                        type="text"
-                        required
-                        className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                      />
-                    </div>
-
-                    <div className="sm:px-3 p-2 mx-2">
-                      <label
-                        htmlFor="keyword"
-                        className=" text-[#2b2b2b] text-sm block "
-                      >
-                        {t('keyword')}
-                      </label>
-                      <FormInput
-                        id="keyword"
-                        onChange={(event) => handleTyping(event)}
-                        name="keyword"
-                        value={productFormValue && productFormValue.keyword}
-                        type="text"
-                        required
-                        className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                      />
-                    </div>
-
-                    <div className="sm:px-3 p-2 mx-2">
-                      <label
-                        htmlFor="expiredAt"
-                        className=" text-[#2b2b2b] text-sm block "
-                      >
-                        {t('expiration_date')}
-                      </label>
-                      <FormInput
-                        id="expiredAt"
-                        onChange={(event) => handleTyping(event)}
-                        name="expiredAt"
-                        value={productFormValue && productFormValue.expiredAt}
-                        type="date"
-                        required
-                        className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                      />
-                    </div>
-                    <div className="sm:px-3 p-2 mx-2">
-                      <label
-                        htmlFor="description"
-                        className=" text-[#2b2b2b] text-sm block "
-                      >
-                        {t('availability')}
-                      </label>
-                      <select
-                        id="status"
-                        name="status"
-                        defaultValue={
-                          productFormValue && productFormValue.status
-                        }
-                        onChange={(event) => handleTyping(event)}
-                        className=" inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                      >
-                        <option value="" selected={!productFormValue}>
-                          {' '}
-                          -- Select --{' '}
-                        </option>
-                        {['AVAILABLE', 'UNAVAILABLE'].map((statusElement) => (
-                          <option key={statusElement} value={statusElement}>
-                            {statusElement}
-                          </option>
-                        ))}
-                      </select>
-                    </div>
-                    <div className="sm:px-3 w-full p-2 mx-2">
-                      <label
-                        htmlFor="description"
-                        className=" text-[#2b2b2b] text-sm block "
-                      >
-                        {t('description')}
-                      </label>
-                      <textarea
-                        id="description"
-                        name="description"
-                        value={productFormValue && productFormValue.description}
-                        onChange={(event) => handleTyping(event)}
-                        className=" mt-2 border-[1px] border-gray-400 w-full rounded-md p-2"
-                        rows={4}
-                      />
-                    </div>
-                    <FormInput
-                      type="hidden"
-                      name="id"
-                      value={productFormValue && productFormValue.id}
-                      onChange={(event) => handleTyping(event)}
-                    />
-                    <button
-                      type="submit"
-                      className="mt-2 w-full py-2 text-xl bg-[#f97316] font-semibold text-white rounded-md active:bg-white active:border-[#f97316] active:text-[#f97316] active:border-[1px]"
-                    >
-                      <span
-                        id="spin"
-                        className={`${
-                          updateProductbtn && updateProductbtn.loading
-                        } h-5 w-5 animate-spin rounded-full border-white border-2 border-solid border-current border-r-transparent motion-reduce:animate-[spin_1.5s_linear_infinite] mr-3 `}
-                        role="status"
-                        disabled={updateProductbtn.disabled}
-                      />
-                      {updateProductbtn && updateProductbtn.text}
-                    </button>
-                  </div>
-                </form>
-                <div />
-              </div>
-              {/* End Product */}
-
-              {/* Start Product Attributes */}
-              <div className="product z-[99] border-2 rounded p-3">
-                <h1 className="z-[9999] font-mono text-lg ">
-                  {t('product_attribute')}{' '}
-                </h1>
-                <hr />
-                {productFormValue?.ProductAttributes ? (
-                  productFormValue?.ProductAttributes.map((productAttibute) => (
-                    <form
-                      key={productAttibute.id}
-                      className=" flex justify-evenly flex-row md:flex-col sm:flex-col md:h-fit items-center sm:items-center py-3 mb-3 border rounded px-2 "
-                      onSubmit={handleUpdateProductAttribute}
-                    >
-                      <div className="w-[33%] md:w-[80%] md:max-h-fit  md:mb-[2%] md md:h-fit relative">
-                        {imgInputController(
-                          imgAttributeDisplay,
-                          productAttibute.attrImage,
-                          false,
-                          t('product_image')
-                        )}
-                      </div>
-                      <div className="relative w-[100%] min-w-[20em flex flex-col  items-end md:items-center ">
-                        <div className=" flex md:flex-col flex-wrap ">
-                          <div className="p-2 w-1/2 sm:w-full relative">
-                            <label
-                              htmlFor="name"
-                              className=" text-[#2b2b2b] text-sm block "
-                            >
-                              {t('attr_name')}
-                            </label>
-                            <FormInput
-                              id="varitationName"
-                              onChange={(event) =>
-                                handleTypingAttribute(event, productAttibute.id)
-                              }
-                              name="varitationName"
-                              value={productAttibute.varitationName}
-                              type="text"
-                              required
-                              className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                            />
-                          </div>
-
-                          <div className="p-2 w-1/2 md:w-full">
-                            <label
-                              htmlFor="quantity"
-                              className=" text-[#2b2b2b] text-sm block "
-                            >
-                              {t('quantity')}
-                            </label>
-                            <FormInput
-                              id="quantity"
-                              onChange={(event) =>
-                                handleTypingAttribute(event, productAttibute.id)
-                              }
-                              name="quantity"
-                              value={productAttibute.quantity}
-                              type="text"
-                              required
-                              className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                            />
-                          </div>
-
-                          <div className="p-2  w-1/2 md:w-full">
-                            <label
-                              htmlFor="price"
-                              className=" text-[#2b2b2b] text-sm block "
-                            >
-                              {t('price')}
-                            </label>
-                            <FormInput
-                              id="price"
-                              onChange={(event) =>
-                                handleTypingAttribute(event, productAttibute.id)
-                              }
-                              name="price"
-                              value={productAttibute.price}
-                              type="text"
-                              required
-                              className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                            />
-                          </div>
-
-                          <div className="p-2  w-1/2 md:w-full">
-                            <label
-                              htmlFor="size"
-                              className=" text-[#2b2b2b] text-sm block "
-                            >
-                              {t('size')}
-                            </label>
-                            <FormInput
-                              id="size"
-                              onChange={(event) =>
-                                handleTypingAttribute(event, productAttibute.id)
-                              }
-                              name="size"
-                              value={productAttibute.size}
-                              type="text"
-                              required
-                              className="inputFormClass placeholder-shown:border-t-blue-gray-200 focus:border-1"
-                            />
-                          </div>
-                          <FormInput
-                            type="hidden"
-                            name="id"
-                            value={productAttibute.id}
-                            onChange={(event) =>
-                              handleTypingAttribute(event, productAttibute.id)
-                            }
-                          />
-                        </div>
-                        <button
-                          type="submit"
-                          className="mt-2 w-2/5 py-1 text-md float-left bg-[#f97316] font-semibold text-white rounded-md active:bg-white active:border-[#f97316] active:text-[#f97316] active:border-[1px]"
-                        >
-                          <span
-                            id="spin"
-                            className={`${
-                              updateProductbtn && updateProductbtn.loading
-                            } h-5 w-5 animate-spin rounded-full border-white border-2 border-solid border-current border-r-transparent motion-reduce:animate-[spin_1.5s_linear_infinite] mr-3 `}
-                            role="status"
-                            disabled={updateProductbtn.disabled}
-                          />
-                          {updateProductbtn && updateProductbtn.text}
-                        </button>
-                      </div>
-                    </form>
-                  ))
-                ) : (
-                  <div> {t('no_data_found')} </div>
-                )}
-                <div />
-              </div>
-              {/* End Product Attributes */}
-            </div>
-          </div>
-        </div>
-      </div>
+      <Modal isModalOpen={isModalOpen} handleCloseModal={handleCloseModal}>
+        {modalProduct && <FullProduct product={modalProduct} />}
+        {modalProduct && <FullProductAttribute product={modalProduct} />}
+      </Modal>
     </div>
   );
 };
